+++ browser-switch/src/main/java/com/braintreepayments/api/BrowserSwitchClient.java
@@ -19,20 +19,20 @@ public class BrowserSwitchClient {
     private final BrowserSwitchInspector browserSwitchInspector;
     private final BrowserSwitchPersistentStore persistentStore;
 
-    private final CustomTabsIntent.Builder customTabsIntentBuilder;
+    private final ChromeCustomTabsInternalClient customTabsInternalClient;
 
     /**
      * Construct a client that manages the logic for browser switching.
      */
     public BrowserSwitchClient() {
-        this(new BrowserSwitchInspector(), BrowserSwitchPersistentStore.getInstance(), new CustomTabsIntent.Builder());
+        this(new BrowserSwitchInspector(), BrowserSwitchPersistentStore.getInstance(), new ChromeCustomTabsInternalClient());
     }
 
     @VisibleForTesting
-    BrowserSwitchClient(BrowserSwitchInspector browserSwitchInspector, BrowserSwitchPersistentStore persistentStore, CustomTabsIntent.Builder customTabsIntentBuilder) {
+    BrowserSwitchClient(BrowserSwitchInspector browserSwitchInspector, BrowserSwitchPersistentStore persistentStore, ChromeCustomTabsInternalClient customTabsInternalClient) {
         this.browserSwitchInspector = browserSwitchInspector;
         this.persistentStore = persistentStore;
-        this.customTabsIntentBuilder = customTabsIntentBuilder;
+        this.customTabsInternalClient = customTabsInternalClient;
     }
 
     /**
@@ -49,14 +49,14 @@ public class BrowserSwitchClient {
 
         Uri browserSwitchUrl = browserSwitchOptions.getUrl();
         int requestCode = browserSwitchOptions.getRequestCode();
+        String returnUrlScheme = browserSwitchOptions.getReturnUrlScheme();
 
         JSONObject metadata = browserSwitchOptions.getMetadata();
         BrowserSwitchRequest request =
-                new BrowserSwitchRequest(requestCode, browserSwitchUrl, metadata);
+                new BrowserSwitchRequest(requestCode, browserSwitchUrl, metadata, returnUrlScheme);
         persistentStore.putActiveRequest(request, appContext);
 
-        CustomTabsIntent customTabsIntent = customTabsIntentBuilder.build();
-        customTabsIntent.launchUrl(activity, browserSwitchUrl);
+        customTabsInternalClient.launchUrl(activity, browserSwitchUrl);
     }
 
     void assertCanPerformBrowserSwitch(FragmentActivity activity, BrowserSwitchOptions browserSwitchOptions) throws BrowserSwitchException {
@@ -120,10 +120,10 @@ public class BrowserSwitchClient {
         BrowserSwitchResult result;
 
         Uri deepLinkUrl = intent.getData();
-        if (deepLinkUrl == null) {
-            result = new BrowserSwitchResult(BrowserSwitchStatus.CANCELED, request);
-        } else {
+        if (deepLinkUrl != null && request.matchesDeepLinkUrlScheme(deepLinkUrl)) {
             result = new BrowserSwitchResult(BrowserSwitchStatus.SUCCESS, request, deepLinkUrl);
+        } else {
+            result = new BrowserSwitchResult(BrowserSwitchStatus.CANCELED, request);
         }
 
         // ensure that browser switch result is delivered exactly one time